# dist/openapi.yaml
...
openapi: 3.0.0
info:
  title: Recipes api
  description:  API for recipes project
  version: "1.0.0"
servers:
- url: "http://127.0.0.1:3000"
#https://virtserver.swaggerhub.com/TalyaSakov/API3.1/1.0.0
  description: SwaggerHub API Auto Mocking hey
...
paths:
...
  /users/add_personal_recipe:
    post:
      tags:
      - user
      summary: logged in user is adding personal recipe
      requestBody:
        content:
          application/json:
              schema:
                $ref: '#/components/schemas/expendedInformation'
      responses:
        "200":
          description: The Recipe successfully saved in your personal recipes list!
        "400":
          description: problem have occured
...
  /users/favorites:
    post:
      tags:
      - user
      summary: save recipe in the favorites list of a user
      description: save recipe in the favorites list of a user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reciepeId:
                  type: string
                  description: reciepe's id which the user saved
                  example: 716429
      responses:
        "200":
          description: The Recipe successfully saved as favorite
        "401":
          description: user is not authorized
    get:
      tags:
      - user
      summary: Return the favorite recipes that were saved by user
      description: Get an array of recipes that were saved by a specific user
      responses:
        "200":
          description: data about the user's favorites recipes returned
        "401":
          description: user is not authorized
...
components:
  schemas:
...
    analyzedInstructions:
      type: object
      properties:
        name:
          type: string
          description: name for the instractions
          example: ""
        steps:
          type: array
          items:
            type: object
            properties:
              name:
                type: integer
                example: 1
              step:
                type: string
                example: wash and rinse pork chops and place into the skillet.cut them into bite sized pieces and add half of the Basil Garlic simmer sauce.boil your water and start working on cooking your bow-tie pasta.when you have finished with boiling and draining your pasta, add it to the pork along with the rest of the Basil Garlic Simmering Sauce, mixing lightly.Next you will top with the Chunky Bruschetta Finishing Sauce, cover with Parmesan, and cover. Cooking on low heat 2 to 3 minutes or until heated through.
              ingredients:
                type: array
                items:
                  $ref: '#/components/schemas/ingredients'
              equipment:
                type: string
                example: 
              length:
                type: string
                example: 
    ingredients:
      type: object
      properties:
        id:
          type: integer
          description: ingredient number
          example: 10120420
        name:
          type: string
          description: ingredient name
          example: farfalle
        localizedName:
          type: string
          example: farfalle
        image:
          type: string
          description: ingredient pic
          example: farfalle.png
    extendedIngredients:
      type: object
      properties:
        id:
          type: integer
          example: 10120420
        aisle:
          type: string
          example: Pasta and Rice
        image:
          type: string
          description: ingredient pic
          example: farfalle.png
        consistency": 
          type: string
          example: SOLID
        name:
          type: string
          example: bow tie pasta
        nameClean:
          type: string
          example: farfalle
        original:
          type: string
          example: 3 cups bow tie pasta",
        originalName:
          type: string
          example: bow tie pasta
        amount: 
          type: integer
          example: 3
        unit:
          type: string
          example: cups
        meta: 
          type: string
        measures:
          type: array
          items:
            type: object
            properties:
              us:
                type: array
                items:
                $ref: '#/components/schemas/measures'
              metric:
                type: array
                items:
                $ref: '#/components/schemas/measures'
    measures:
      type: object
      properties:
        amount:
          type: integer
          example: 1
        unitShort:
          type: string
          example: servings
        unitLong:
          type: string
          example: servings
...
    expendedInformation:
      type: object
      properties:
        title:
          type: string
          description: recipe name
          example: What to make for dinner tonight?? Bruschetta Style Pork & Pasta
        readyInMinutes:
          type: integer
          description: the time takes to make this recipe
          example: 35
        image:
          type: string
          description: recipe image
          format: URL
          example: https://spoonacular.com/recipeImages/715538-556x370.jpg
        aggregateLikes:
          type: integer
          description: how many likes got this recipe
          example: 163
        vegan:
          type: string
          description: indication if the recipe is vegan
          example: false
        vegetarian:
          type: string
          description: indicate if the recipe is vegetarian
          example: false
        glutenFree:
          type: string
          description: indication if the recipe is gluten free
          example: false
        servings:
          type: integer
          description: indicate how many people can eat the recipe
          example: 5
        analyzedInstructions:
          type: array
          items:
            $ref: '#/components/schemas/analyzedInstructions'
        extendedIngredients:
          type: array
          items:
            $ref: '#/components/schemas/extendedIngredients'